Refactor deployment processing in own plugin
refs: #2
refs: #14
