+++ components/bpmn-q/karma.conf.js
@@ -17,12 +17,14 @@ module.exports = function (config) {
             'test/tests/editor/editor.spec.js',
             'test/tests/editor/plugin.spec.js',
             'test/tests/planqk/planqk-transformation.spec.js',
-            'test/tests/quantme/quantme-transformation.spec.js',
             'test/tests/editor/utils/modelling-util.spec.js',
             'test/tests/qhana/qhana-plugin-config.spec.js',
             'test/tests/qhana/qhana-service-configs.spec.js',
+            'test/tests/quantme/quantme-transformation.spec.js',
             'test/tests/quantme/data-object-configs.spec.js',
             'test/tests/quantme/quantme-config.spec.js',
+            'test/tests/opentosca/opentosca-config.spec.js',
+            'test/tests/opentosca/deployment-utils.spec.js',
             'test/tests/dataflow/data-flow-transformation.spec.js',
             'test/tests/dataflow/data-flow-plugin-config.spec.js',
             'test/tests/dataflow/data-flow-configurations-endpoint.spec.js',
