+++ components/bpmn-q/modeler-component/extensions/opentosca/configTabs/OpenToscaTab.js
