+++ components/bpmn-q/modeler-component/extensions/opentosca/deployment/BindingUtils.js
