+++ components/bpmn-q/modeler-component/extensions/opentosca/framework-config/index.js
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) 2021 Institute of Architecture of Application Systems -
+ * University of Stuttgart
+ *
+ * This program and the accompanying materials are made available under the
+ * terms the Apache Software License 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+import * as configManager from './config-manager';
+
+const config = configManager;
+export default config;
