+++ components/bpmn-q/modeler-component/extensions/opentosca/modeling/properties-provider/DmnImplementationProps.js
@@ -8,8 +8,8 @@ import {
     SelectEntry,
     isSelectEntryEdited
 } from '@bpmn-io/properties-panel';
-import {getImplementationType} from "../../../utilities/ImplementationTypeHelperExtension";
 import {useService} from "bpmn-js-properties-panel";
+import {getImplementationType} from "../../../quantme/utilities/ImplementationTypeHelperExtension";
 
 export function DmnImplementationProps(props) {
     const {
