+++ components/bpmn-q/modeler-component/extensions/quantme/utilities/ImplementationTypeHelperExtension.js
@@ -65,7 +65,7 @@ export function getImplementationType(element) {
         return 'delegateExpression';
     }
 
-    const deploymentModelUrl = businessObject.get('quantme:deploymentModelUrl');
+    const deploymentModelUrl = businessObject.get('opentosca:deploymentModelUrl');
     if (typeof deploymentModelUrl !== 'undefined') {
         return 'deploymentModel';
     }
