+++ components/bpmn-q/public/index.html
@@ -129,6 +129,9 @@
             name: 'qhana',
         },
         {
+            name: 'opentosca',
+        },
+        {
             name: 'planqk',
             config: {
                 serviceEndpointBaseUrl: 'http://dummy.com',
