+++ components/bpmn-q/test/tests/quantme/quantme-config.spec.js
@@ -20,8 +20,6 @@ describe('Test QuantME ConfigManager', function () {
                     name: 'quantme',
                     config: {
                         quantmeDataConfigurationsEndpoint: 'http://test:8100/data-objects',
-                        opentoscaEndpoint: 'http://test:1337/csars',
-                        wineryEndpoint: 'http://test:8093/winery',
                         nisqAnalyzerEndpoint: 'http://test:8098/nisq-analyzer',
                         transformationFrameworkEndpoint: 'http://test:8888',
                         qiskitRuntimeHandlerEndpoint: 'http://test:8889',
@@ -37,8 +35,6 @@ describe('Test QuantME ConfigManager', function () {
             );
 
             expect(quantmeConfig.getQuantMEDataConfigurationsEndpoint()).to.equal('http://test:8100/data-objects');
-            expect(quantmeConfig.getOpenTOSCAEndpoint()).to.equal('http://test:1337/csars');
-            expect(quantmeConfig.getWineryEndpoint()).to.equal('http://test:8093/winery');
             expect(quantmeConfig.getNisqAnalyzerEndpoint()).to.equal('http://test:8098/nisq-analyzer');
             expect(quantmeConfig.getTransformationFrameworkEndpoint()).to.equal('http://test:8888');
             expect(quantmeConfig.getQiskitRuntimeHandlerEndpoint()).to.equal('http://test:8889');
