+++ components/bpmn-q/webpack.config.js
@@ -69,8 +69,9 @@ module.exports = {
             ENABLE_PLANQK_PLUGIN: true,
             ENABLE_QHANA_PLUGIN: true,
             ENABLE_QUANTME_PLUGIN: true,
-            GITHUB_TOKEN: '',
+            ENABLE_OPENTOSCA_PLUGIN: true,
             OPENTOSCA_ENDPOINT: 'http://localhost:1337/csars',
+            GITHUB_TOKEN: '',
             NISQ_ANALYZER_ENDPOINT: 'http://localhost:8098/nisq-analyzer',
             QISKIT_RUNTIME_HANDLER_ENDPOINT: 'http://localhost:8889',
             QHANA_GET_PLUGIN_URL: 'http://localhost:5006/api/plugins/',
